refact: move TextField to platform.theme.ui.input
