+++ feature/auth/src/main/java/com/jeanbarrossilva/orca/feature/auth/Auth.kt
@@ -32,7 +32,7 @@ import com.jeanbarrossilva.orca.core.feed.profile.account.Account
 import com.jeanbarrossilva.orca.core.sample.feed.profile.account.sample
 import com.jeanbarrossilva.orca.platform.theme.OrcaTheme
 import com.jeanbarrossilva.orca.platform.theme.ui.action.PrimaryButton
-import com.jeanbarrossilva.orca.platform.ui.component.input.TextField
+import com.jeanbarrossilva.orca.platform.theme.ui.input.TextField
 import com.jeanbarrossilva.orca.platform.ui.core.requestFocusWithDelay
 
 /** Tag that identifies the username field for testing purposes. **/
