+++ feature/search/src/main/java/com/jeanbarrossilva/orca/feature/search/Search.kt
@@ -44,7 +44,7 @@ import com.jeanbarrossilva.orca.core.sample.feed.profile.search.samples
 import com.jeanbarrossilva.orca.feature.search.ui.SearchResultCard
 import com.jeanbarrossilva.orca.platform.theme.OrcaTheme
 import com.jeanbarrossilva.orca.platform.theme.extensions.backwardsNavigationArrow
-import com.jeanbarrossilva.orca.platform.ui.component.input.TextField
+import com.jeanbarrossilva.orca.platform.theme.ui.input.TextField
 import com.jeanbarrossilva.orca.platform.ui.core.requestFocusWithDelay
 
 internal object SearchDefaults {
